+++ brane-ast/src/traversals/prune.rs
@@ -1,29 +1,30 @@
 //  RETURN.rs
 //    by Lut99
-// 
+//
 //  Created:
 //    31 Aug 2022, 18:00:09
 //  Last edited:
-//    23 Dec 2022, 16:36:08
+//    02 Nov 2023, 14:12:25
 //  Auto updated?
 //    Yes
-// 
+//
 //  Description:
 //!   Traversal that prunes the AST for compilation.
-//! 
+//!
 //!   In particular, inserts return statements into functions such that there
 //!   if one for every codepath and compiles for-loops to while-statements.
-// 
+//
 
 use std::cell::Ref;
+use std::collections::HashSet;
 use std::mem;
 
-use brane_dsl::{DataType, TextPos, TextRange};
-use brane_dsl::symbol_table::FunctionEntry;
 use brane_dsl::ast::{Block, Node, Program, Stmt};
+use brane_dsl::symbol_table::FunctionEntry;
+use brane_dsl::{DataType, TextPos, TextRange};
 
-pub use crate::errors::PruneError as Error;
 use crate::errors::AstError;
+pub use crate::errors::PruneError as Error;
 
 
 /***** TESTS *****/
@@ -33,8 +34,9 @@ mod tests {
     use brane_shr::utilities::{create_data_index, create_package_index, test_on_dsl_files};
     use specifications::data::DataIndex;
     use specifications::package::PackageIndex;
-    use super::*;
+
     use super::super::print::dsl;
+    use super::*;
     use crate::{compile_program_to, CompileResult, CompileStage};
 
 
@@ -48,7 +50,7 @@ mod tests {
 
             // Load the package index
             let pindex: PackageIndex = create_package_index();
-            let dindex: DataIndex    = create_data_index();
+            let dindex: DataIndex = create_data_index();
 
             // Run up to this traversal
             let program: Program = match compile_program_to(code.as_bytes(), &pindex, &dindex, &ParserOptions::bscript(), CompileStage::Prune) {
@@ -63,7 +65,7 @@ mod tests {
                     // Print the error
                     err.prettyprint(path.to_string_lossy(), &code);
                     panic!("Failed to prune AST (see output above)");
-                }
+                },
                 CompileResult::Err(errs) => {
                     // Print the errors
                     for e in errs {
@@ -72,7 +74,9 @@ mod tests {
                     panic!("Failed to prune AST (see output above)");
                 },
 
-                _ => { unreachable!(); },
+                _ => {
+                    unreachable!();
+                },
             };
 
             // Now print the file for prettyness
@@ -88,22 +92,22 @@ mod tests {
 
 /***** TRAVERSAL FUNCTIONS *****/
 /// Prunes the statements in the given block for compilation.
-/// 
+///
 /// # Arguments
 /// - `block`: The Block to prune.
-/// 
+///
 /// # Returns
 /// Whether or not the block completely returns or not. Also alters, adds or removes statements to or from the block.
-/// 
+///
 /// # Errors
 /// This function may error if a given statement in the block is a function that does not correctly return on all paths.
-/// 
+///
 /// If an error occurred, it is written to the given `errors` list. The function then still returns whether this block itself fully returns or not.
 fn pass_block(block: &mut Block, errors: &mut Vec<Error>) -> bool {
     // Iterate over the statements in the block.
-    let old_stmts         : Vec<Stmt> = mem::take(&mut block.stmts);
-    let mut new_stmts     : Vec<Stmt> = Vec::with_capacity(old_stmts.len());
-    let mut fully_returns : bool      = false;
+    let old_stmts: Vec<Stmt> = mem::take(&mut block.stmts);
+    let mut new_stmts: Vec<Stmt> = Vec::with_capacity(old_stmts.len());
+    let mut fully_returns: bool = false;
     for s in old_stmts {
         // Run 'em through the statements (to replace for's into while's and such)
         let (mut new_stmt, returns): (Vec<Stmt>, bool) = pass_stmt(s, errors);
@@ -124,17 +128,17 @@ fn pass_block(block: &mut Block, errors: &mut Vec<Error>) -> bool {
 }
 
 /// Prunes the given statement for compilation.
-/// 
+///
 /// # Arguments
 /// - `stmt`: The statement to prune.
 /// - `errors`: The list that can keep track of multiple errors.
-/// 
+///
 /// # Returns
 /// A tuple of a (series of) Stmt(s) to replace the given one, and whether this statement _fully_ returns. This list will typically be the given statement only, but not necessarily so.
-/// 
+///
 /// # Errors
 /// This function may error if the given statement is a function that does not correctly return on all paths.
-/// 
+///
 /// If an error occurred, it is written to the given `errors` list. The function then still returns whether this statement fully returns or not.
 fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
     let mut stmt: Stmt = stmt;
@@ -147,7 +151,7 @@ fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
             let returns: bool = pass_block(block, errors);
 
             // Return the statement as-is
-            (vec![ stmt ], returns)
+            (vec![stmt], returns)
         },
 
         FuncDef { code, st_entry, .. } => {
@@ -162,52 +166,53 @@ fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
                     e.signature.ret.clone()
                 };
                 if ret_type != DataType::Any && ret_type != DataType::Void {
-                    errors.push(Error::MissingReturn { expected: ret_type, range: TextRange::new(TextPos::new(code.end().line, code.end().col - 1), code.end().clone()) });
-                    return (vec![ stmt ], false);
+                    errors.push(Error::MissingReturn {
+                        expected: ret_type,
+                        range:    TextRange::new(TextPos::new(code.end().line, code.end().col - 1), code.end().clone()),
+                    });
+                    return (vec![stmt], false);
                 }
 
                 // Otherwise, insert a void return
-                code.stmts.push(Stmt::Return{ expr: None, data_type: ret_type, range: TextRange::none() });
+                code.stmts.push(Stmt::Return { expr: None, data_type: ret_type, output: HashSet::new(), range: TextRange::none() });
             }
 
             // Done (the function definition itself never returns)
-            (vec![ stmt ], false)
+            (vec![stmt], false)
         },
-        ClassDef{ methods, .. } => {
+        ClassDef { methods, .. } => {
             // Recurse into all of the methods
             for m in methods {
                 let old_m: Stmt = mem::take(m);
                 let (mut new_m, _) = pass_stmt(old_m, errors);
-                if new_m.len() != 1 { panic!("Method statement was pruned to something else than 1 statement; this should never happen!"); }
+                if new_m.len() != 1 {
+                    panic!("Method statement was pruned to something else than 1 statement; this should never happen!");
+                }
                 *m = Box::new(new_m.pop().unwrap());
             }
             // The class definition itself never returns
-            (vec![ stmt ], false)
+            (vec![stmt], false)
         },
-        Return{ .. } => {
+        Return { .. } => {
             // Clearly, a return statement always returns
-            (vec![ stmt ], true)
+            (vec![stmt], true)
         },
 
-        If{ consequent, alternative, .. } => {
+        If { consequent, alternative, .. } => {
             // Inspect if the consequent fully returns
             let true_returns: bool = pass_block(consequent, errors);
             // Inspect if the alternative returns
-            let false_returns: bool = if let Some(alternative) = alternative {
-                pass_block(alternative, errors)
-            } else {
-                false
-            };
+            let false_returns: bool = if let Some(alternative) = alternative { pass_block(alternative, errors) } else { false };
 
             // This if-statement returns if both blocks return
-            (vec![ stmt ], true_returns && false_returns)
+            (vec![stmt], true_returns && false_returns)
         },
-        For{ initializer, condition, increment, consequent, range, .. } => {
-            let initializer : Stmt                      = mem::take(initializer);
-            let condition   : brane_dsl::ast::Expr      = mem::take(condition);
-            let increment   : Stmt                      = mem::take(increment);
-            let mut consequent  : brane_dsl::ast::Block = mem::take(consequent);
-            let range       : TextRange                 = mem::take(range);
+        For { initializer, condition, increment, consequent, range, .. } => {
+            let initializer: Stmt = mem::take(initializer);
+            let condition: brane_dsl::ast::Expr = mem::take(condition);
+            let increment: Stmt = mem::take(increment);
+            let mut consequent: brane_dsl::ast::Block = mem::take(consequent);
+            let range: TextRange = mem::take(range);
 
             // We transform this for-loop to a while-loop first
 
@@ -219,11 +224,7 @@ fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
             consequent.stmts.push(increment);
 
             // Step 3: Write the condition + updated consequent as a new While loop
-            let while_stmt: Stmt = Stmt::While {
-                condition,
-                consequent : Box::new(consequent),
-                range,
-            };
+            let while_stmt: Stmt = Stmt::While { condition, consequent: Box::new(consequent), range };
 
             // Step 4: Analyse as a normal while-loop (increment is not (yet) needed here)
             let (mut while_stmt, returns): (Vec<Stmt>, bool) = pass_stmt(while_stmt, errors);
@@ -232,31 +233,33 @@ fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
             // Step 5: Done
             (stmts, returns)
         },
-        While{ consequent, .. } => {
+        While { consequent, .. } => {
             // Check if the block returns
             let returns: bool = pass_block(consequent, errors);
-            (vec![ stmt ], returns)
+            (vec![stmt], returns)
         },
-        On{ block, .. } => {
+        On { block, .. } => {
             // Check if the block returns
             let returns: bool = pass_block(block, errors);
-            (vec![ stmt ], returns)
+            (vec![stmt], returns)
         },
-        Parallel{ blocks, .. } => {
+        Parallel { blocks, .. } => {
             // A Parallel statement cannot return, but technically might define functions to still recurse
             for b in blocks {
                 let old_b: Stmt = mem::take(b);
                 let (mut new_b, _) = pass_stmt(old_b, errors);
-                if new_b.len() != 1 { panic!("Parallel block statement was pruned to something else than 1 statement; this should never happen!"); }
+                if new_b.len() != 1 {
+                    panic!("Parallel block statement was pruned to something else than 1 statement; this should never happen!");
+                }
                 *b = Box::new(new_b.pop().unwrap());
             }
 
             // Done
-            (vec![ stmt ], false)
+            (vec![stmt], false)
         },
 
         // The rest we don't care about in this traversal
-        _ => { (vec![ stmt ], false) }
+        _ => (vec![stmt], false),
     }
 }
 
@@ -266,15 +269,15 @@ fn pass_stmt(stmt: Stmt, errors: &mut Vec<Error>) -> (Vec<Stmt>, bool) {
 
 /***** LIBRARY *****/
 /// Prunes the given `brane-dsl` AST for compilation.
-/// 
+///
 /// Note that the previous traversals should all already have come to pass.
-/// 
+///
 /// # Arguments
 /// - `root`: The root node of the tree on which this compiler pass will be done.
-/// 
+///
 /// # Returns
 /// The same nodes as went in, but now ready for compilation.
-/// 
+///
 /// # Errors
 /// This pass may throw multiple `AstError::PruneErrors`s if the locations could not be satisactorily deduced.
 pub fn do_traversal(root: Program) -> Result<Program, Vec<AstError>> {
@@ -285,9 +288,5 @@ pub fn do_traversal(root: Program) -> Result<Program, Vec<AstError>> {
     pass_block(&mut root.block, &mut errors);
 
     // Done
-    if errors.is_empty() {
-        Ok(root)
-    } else {
-        Err(errors.into_iter().map(|e| e.into()).collect())
-    }
+    if errors.is_empty() { Ok(root) } else { Err(errors.into_iter().map(|e| e.into()).collect()) }
 }